Moved ShutdownManager into an internal package.
This meant having to also move the shutdownManager property into the interal driver DSL interface, which further meant also moving RPCDriver into an internal package (which makes sense as it's not a public testing API).
