+++ testing/node-driver/src/main/kotlin/net/corda/testing/driver/Driver.kt
@@ -37,6 +37,7 @@ import net.corda.nodeapi.internal.addShutdownHook
 import net.corda.testing.*
 import net.corda.testing.common.internal.NetworkParametersCopier
 import net.corda.testing.common.internal.testNetworkParameters
+import net.corda.testing.internal.ShutdownManager
 import net.corda.testing.node.ClusterSpec
 import net.corda.testing.node.MockServices.Companion.MOCK_VERSION_INFO
 import net.corda.testing.node.NotarySpec
@@ -54,12 +55,10 @@ import java.time.Instant
 import java.time.ZoneOffset.UTC
 import java.time.format.DateTimeFormatter
 import java.util.*
-import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
 import java.util.concurrent.ScheduledExecutorService
 import java.util.concurrent.TimeUnit.MILLISECONDS
 import java.util.concurrent.TimeUnit.SECONDS
-import java.util.concurrent.TimeoutException
 import java.util.concurrent.atomic.AtomicInteger
 import kotlin.collections.ArrayList
 import kotlin.concurrent.thread
@@ -173,6 +172,14 @@ interface DriverDSLExposedInterface : CordformContext {
     fun startWebserver(handle: NodeHandle, maximumHeapSize: String): CordaFuture<WebserverHandle>
 
     fun waitForAllNodesToFinish()
+}
+
+interface DriverDSLInternalInterface : DriverDSLExposedInterface {
+    val shutdownManager: ShutdownManager
+
+    fun start()
+
+    fun shutdown()
 
     /**
      * Polls a function until it returns a non-null value. Note that there is no timeout on the polling.
@@ -192,13 +199,6 @@ interface DriverDSLExposedInterface : CordformContext {
     fun pollUntilTrue(pollName: String, pollInterval: Duration = DEFAULT_POLL_INTERVAL, warnCount: Int = DEFAULT_WARN_COUNT, check: () -> Boolean): CordaFuture<Unit> {
         return pollUntilNonNull(pollName, pollInterval, warnCount) { if (check()) Unit else null }
     }
-
-    val shutdownManager: ShutdownManager
-}
-
-interface DriverDSLInternalInterface : DriverDSLExposedInterface {
-    fun start()
-    fun shutdown()
 }
 
 sealed class NodeHandle {
@@ -557,102 +557,6 @@ fun <A> poll(
     }
     executorService.submit(task) // The check may be expensive, so always run it in the background even the first time.
     return resultFuture
-}
-
-class ShutdownManager(private val executorService: ExecutorService) {
-    private class State {
-        val registeredShutdowns = ArrayList<CordaFuture<() -> Unit>>()
-        var isShutdown = false
-    }
-
-    private val state = ThreadBox(State())
-
-    companion object {
-        inline fun <A> run(providedExecutorService: ExecutorService? = null, block: ShutdownManager.() -> A): A {
-            val executorService = providedExecutorService ?: Executors.newScheduledThreadPool(1)
-            val shutdownManager = ShutdownManager(executorService)
-            try {
-                return block(shutdownManager)
-            } finally {
-                shutdownManager.shutdown()
-                providedExecutorService ?: executorService.shutdown()
-            }
-        }
-    }
-
-    fun shutdown() {
-        val shutdownActionFutures = state.locked {
-            if (isShutdown) {
-                emptyList<CordaFuture<() -> Unit>>()
-            } else {
-                isShutdown = true
-                registeredShutdowns
-            }
-        }
-        val shutdowns = shutdownActionFutures.map { Try.on { it.getOrThrow(1.seconds) } }
-        shutdowns.reversed().forEach {
-            when (it) {
-                is Try.Success ->
-                    try {
-                        it.value()
-                    } catch (t: Throwable) {
-                        log.warn("Exception while shutting down", t)
-                    }
-                is Try.Failure -> log.warn("Exception while getting shutdown method, disregarding", it.exception)
-            }
-        }
-    }
-
-    fun registerShutdown(shutdown: CordaFuture<() -> Unit>) {
-        state.locked {
-            require(!isShutdown)
-            registeredShutdowns.add(shutdown)
-        }
-    }
-
-    fun registerShutdown(shutdown: () -> Unit) = registerShutdown(doneFuture(shutdown))
-
-    fun registerProcessShutdown(processFuture: CordaFuture<Process>) {
-        val processShutdown = processFuture.map { process ->
-            {
-                process.destroy()
-                /** Wait 5 seconds, then [Process.destroyForcibly] */
-                val finishedFuture = executorService.submit {
-                    process.waitFor()
-                }
-                try {
-                    finishedFuture.get(5, SECONDS)
-                } catch (exception: TimeoutException) {
-                    finishedFuture.cancel(true)
-                    process.destroyForcibly()
-                }
-                Unit
-            }
-        }
-        registerShutdown(processShutdown)
-    }
-
-    interface Follower {
-        fun unfollow()
-        fun shutdown()
-    }
-
-    fun follower() = object : Follower {
-        private val start = state.locked { registeredShutdowns.size }
-        private val end = AtomicInteger(start - 1)
-        override fun unfollow() = end.set(state.locked { registeredShutdowns.size })
-        override fun shutdown() = end.get().let { end ->
-            start > end && throw IllegalStateException("You haven't called unfollow.")
-            state.locked {
-                registeredShutdowns.subList(start, end).listIterator(end - start).run {
-                    while (hasPrevious()) {
-                        previous().getOrThrow().invoke()
-                        set(doneFuture {}) // Don't break other followers by doing a remove.
-                    }
-                }
-            }
-        }
-    }
 }
 
 class DriverDSL(
